pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "institutional-investor-casebook"
version = "0.1.0"
description = "Bridging hedge fund domain expertise with LLM evaluation pipelines."
authors = [
    {name = "Brad Schonhoft", email = "brads@pmtoaillc.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "accelerate>=0.29.0",
    "bitsandbytes>=0.43.0",
    "python-dotenv>=1.0.0",
    "huggingface_hub>=0.22.0",
    "sentence-transformers>=2.7.0", # For future scoring logic
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]